+++ src/tool/xmeta/cpp/models/namespace_member_model.cpp
@@ -1,0 +1,14 @@
+#include "namespace_member_model.h"
+#include "namespace_model.h"
+
+namespace xlang::xmeta
+{
+    std::string const namespace_member_model::get_fully_qualified_id() const noexcept
+    {
+        if (containing_namespace_name)
+        {
+            return containing_namespace_name.value() + "." + get_id();
+        }
+        return get_containing_namespace_body()->get_containing_namespace()->get_fully_qualified_id() + "." + get_id();
+    }
+}
